\name{Description}
\alias{Description}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{first_r_package
%%  ddf_pr
%%  calc_ddf
%%  spatial_data_read
%%  forcing_dates
%%  data_read

}
\description{
%%  ddf_pr is used for precipitation ddf (depth duration frequency curves)
%%  calc_ddf calculates ddf
%%  spatial_data_read is able to read spatial data and process
%%  forcing_dates is used to get dates from files names
%%  data_read will read in data
}
\usage{
Description(x)
}
\author{
%% Name:D
%% Email:jupsijups888@gmail.com
\note{
%% ~~Any questions email me~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}

% \keyword{Hydrology}
% \keyword{R}
% \keyword{GIS}
% \keyword{DDF}
% \concept{Data Science}
% \concept{Spatial Data}
% \concept{Environmental Data Science}

