% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorrs.R
\name{ggCorrs}
\alias{ggCorrs}
\title{ACF, PACF and CCF}
\usage{
ggCorrs(
  x,
  y = NULL,
  min_lag = -20,
  max_lag = 20,
  plot_type = "ACF",
  print_vals = F,
  alpha = 0.05
)
}
\arguments{
\item{x, y}{Variables of interest.}

\item{min_lag, max_lag}{They correspond to the minimum lag and the maximum lag. Default values are \code{min_lag=-20} and \code{max_lag=20}.}

\item{plot_type}{It specifies the type of the plot. You can choose \code{"ACF"} (default) for the autocorrelation function, \code{"PACF"} for the partial autocorrelation function and \code{"CCF"} for the cross-correlation function.}

\item{alpha}{The type-I error to build confidence intervals. The default value is \code{alpha=.05}.}
}
\value{
Return the desired plot. It is possible to extract the lags and the corresponding values.

The confidence intervals are calculated using the asymptotic distribution of the correlation estimator \eqn{\rho}~N(0,T^\eqn{{-1}}). Therefore IC_alpha(\eqn{\rho})=(-qnorm(1-alpha/2)sqrt(T^\eqn{{-1}}), +qnorm(1-alpha/2)sqrt(T^\eqn{{-1}})).
}
\description{
It uses the \pkg{ggplot2} package to create three usefull functions in time series analysis: ACF, PACF and CCF.
}
\examples{
# ::::::::::: #
# TOY EXAMPLE #
# ::::::::::: #

library("ggplot2")
set.seed(111)
x <- rnorm(50)
y <- rchisq(50, 4)
g <- ggCorrs(x)
# g <- ggCorrs(x, print_vals = T)
g$data$lag # printing lags
g$data$acf.val # printing values of ACF
ggCorrs(x, y, plot_type = "CCF")


}
